Moved Navbar. Split NavBar into parts. NavBar balance now calls backend query.
