+++ src/components/Profile/ProfilePage.jsx
@@ -1,5 +1,5 @@
 import React from 'react';
-import Navbar from '../Navbar';
+import Navbar from '../Shared/Navbar';
 import Profile from './Profile';
 
 const ProfilePage = () => {
